// src/enums/settings/theme-enum.ts
/**
 * 主题枚举
 */
export const enum ThemeMode {
  /**
   * 明亮主题
   */
  LIGHT = "light",
  /**
   * 暗黑主题
   */
  DARK = "dark",

  /**
   * 系统自动
   */
  AUTO = "auto",
}

/**
 * 侧边栏配色方案枚举
 */
export const enum SidebarColor {
  /**
   * 经典蓝
   */
  CLASSIC_BLUE = "classic-blue",
  /**
   * 极简白
   */
  MINIMAL_WHITE = "minimal-white",
}